pyproject.toml
[tool.poetry]
name = "ptodnes"
version = "1.11.2.4"
description = ""
authors = ["Ondrej Dohnal <xdohna45@vutbr.cz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python  = ">3.11, <=3.15"
aiodns  = "^3.2.0"
aiohttp = "^3.11.13"
aiopg   = "^1.4.0"
pyyaml  = "^6.0.2"
ptlibs  = "^1.0.17"
aiofiles = "^25.1.0"
punycode = "^0.2.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.test]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
sphinx = "^8.2.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"
pylint = "^3.3.1"
pytest-md-report = "^0.6.2"

[tool.poetry.scripts]
ptodnes = "ptodnes.__main__:__main__"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
